backend/requirements.txt
# Database
supabase==2.1.0

# Core Framework
pydantic>=2.6.0
docling>=1.0.0

# Core AI/ML Libraries
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-huggingface>=0.0.1
langchain-text-splitters>=0.0.1
openai>=1.0.0

# Vector Database
qdrant-client>=1.9.0,<2.0.0
langchain-qdrant>=0.1.0

# PDF Processing
pypdf>=3.17.0
pdfplumber>=0.10.0
pymupdf>=1.23.0

# Text Processing
tiktoken>=0.5.0
sentence-transformers>=2.6.0

# Web Framework
fastapi>=0.100.0
uvicorn>=0.23.0

# Environment & Utils
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
aiofiles>=23.2.1

# File Upload
python-multipart>=0.0.6

# system without OpenAI
scikit-learn>=1.3.0